requirements.txt
astroid==2.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
codecov==2.0.15
coverage==4.5.1
decorator==4.3.0
Django==2.1.1
django-appconf==1.0.2
django-datetime-utc==1.0.2
django-nose==1.4.6
django-responsive2==0.1.3
django-tinymce4-lite==1.7.4
django-widget-tweaks==1.4.3
Faker==0.9.2
flake8==3.6.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
idna==2.7
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
jsmin==2.2.2
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
nose==1.3.7
parso==0.3.1
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==2.0.0
Pygments==2.3.1
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.5
rcssmin==1.0.6
requests==2.20.0
rjsmin==1.0.12
selenium==3.14.1
six==1.11.0
snowballstemmer==1.2.1
stripe==2.17.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
wrapt==1.10.11
